adding navbar and adding the images and the fav button to my new database
