+++ .env
@@ -1,3 +1,3 @@
 REACT_APP_KEY=e1d6982df30fca741085187438a1296d
 REACT_APP_URL=https://api.themoviedb.org/3
-REACT_APP_DATABASE=https://movie-library-nx1r.onrender.com/
+REACT_APP_DATABASE=https://moviesapi-5cja.onrender.com/
