+++ src/App.css
@@ -1,0 +1,3 @@
+.main__home{
+    background-image: url('/public/moviesImg.jpg');
+}
