+++ src/App.js
@@ -1,10 +1,17 @@
+import {Routes,Route,BrowserRouter} from 'react-router-dom'
+import Navbar from './Navbar';
 import './App.css';
 import Home from './Home';
+import MovieList from './MovieList';
 function App() {
   return (
-    <div className="App">
-      <Home/>
-    </div>
+    <BrowserRouter className="App">
+      <Navbar/>
+      <Routes>
+        <Route path="/" element={<Home/>}></Route>
+        <Route path="/movieList" element={<MovieList/>}></Route>
+      </Routes>
+    </BrowserRouter>
   );
 }
