+++ src/Home.js
@@ -1,29 +1,10 @@
 import React, { useEffect, useState } from 'react'
-import axios from 'axios'
+import '../src/App.css'
 import MovieList from './MovieList'
-import ModalMovie from './ModalMovie';
-import Row from 'react-bootstrap/Row';
-
 function Home() {
-  const [trendingMovies,setTrendingMovies]=useState([]);
-  const [targetMovie,setTargetMovie]=useState({});
-  const [show, setShow] = useState(false);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
-  
-  console.log(process.env)
-  function getTrending(){
-axios.get(`${process.env.REACT_APP_URL}/trending/all/week?api_key=${process.env.REACT_APP_KEY}&language=end-US`)
-.then(res=>setTrendingMovies(res.data.results))
-.catch(err=>console.log(err));
-  }
-  useEffect(()=>{
-    getTrending()
-  },[])
   return (
-    <div>
-      <MovieList movies={trendingMovies} setShow={setShow} setTargetMovie={setTargetMovie}/>
-      <ModalMovie handleClose={handleClose} handleShow={handleShow} show={show} targetMovie={targetMovie}/>
+    <div className='w-[100%] h-[100vh] flex items-center justify-center'>
+      <MovieList/>
     </div>
   )
 }
