+++ src/ModalMovie.js
@@ -9,8 +9,14 @@ import Form from 'react-bootstrap/Form';
 function ModalMovie({targetMovie,handleClose,handleShow,show}) {
 const [inputValue,setInputValue]=useState("")
 function handleAddToDatabase(){
+  let sentData={
+    title:targetMovie.title,
+    comment:inputValue,
+    relase_date:targetMovie.release_date,
+    rating:Math.floor(targetMovie.vote_average)
+  }
   handleClose();
-    axios.post(`${process.env.REACT_APP_DATABASE}`,inputValue)
+    axios.post(`${process.env.REACT_APP_DATABASE}movies`,sentData)
     .then(res=>console.log(res))
     .catch(err=>console.log(err))
   };
@@ -23,13 +29,14 @@ function handleAddToDatabase(){
    
    <Modal show={show} onHide={handleClose}>
         <Modal.Header closeButton>
-          <Modal.Title>{targetMovie.title}</Modal.Title>
+          <img className='w-[100%] h-[400px]' src={`http://image.tmdb.org/t/p/w500/${targetMovie.poster_path}`} alt="movie IMG" />
         </Modal.Header>
         <Modal.Body>
+        <Modal.Title>{targetMovie.title}</Modal.Title>
           {targetMovie.overview}
         </Modal.Body>
         <InputGroup className="mb-3">
-        <InputGroup.Text id="inputGroup-sizing-default"value={(e)=>setInputValue(e.target.value)}
+        <InputGroup.Text id="inputGroup-sizing-default" onChange={(e)=>setInputValue(e.target.value)}
 >
          Comment
         </InputGroup.Text>
