+++ tailwind.config.js
@@ -5,7 +5,9 @@ module.exports = {
 
   ],
   theme: {
-    extend: {},
+    extend: {
+     
+    },
   },
   plugins: [],
 }
